refactor: move windows to submodules
