+++ source/windows/main_window/__init__.py
@@ -1,0 +1,3 @@
+from .window import BlenderLauncher
+
+__all__ = ["BlenderLauncher"]
