+++ source/windows/onboarding_window/__init__.py
@@ -1,0 +1,3 @@
+from .window import OnboardingWindow
+
+__all__ = ["OnboardingWindow"]
