+++ source/windows/settings_window/__init__.py
@@ -1,0 +1,3 @@
+from .window import SettingsWindow
+
+__all__ = ["SettingsWindow"]
