+++ source/windows/settings_window/connection_tab.py
@@ -30,8 +30,9 @@ from PySide6.QtWidgets import (
     QLineEdit,
     QPushButton,
 )
-from widgets.settings_form_widget import SettingsFormWidget
 from windows.popup_window import Popup
+
+from .settings_form_widget import SettingsFormWidget
 
 if TYPE_CHECKING:
     from windows.main_window import BlenderLauncher
